requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
wget==3.2
